requirements.txt
certifi==2025.8.3
numpy==2.3.2
pandas==2.3.1
polygon-api-client==1.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
tenacity==9.1.2
tqdm==4.67.1
tzdata==2025.2
urllib3==2.5.0
websockets==14.2
